requirement_npu.txt
./localwheels/torch_npu-2.1.0.post7-cp310-cp310-manylinux_2_17_aarch64.manylinux2014_aarch64.whl
torchvision == 0.16.0
numpy
Pillow
nibabel
flask[async]